tests/wpt/tests/html/canvas/element/layers/2d.layer.malformed-operations.getImageData.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.layer.malformed-operations.getImageData</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.layer.malformed-operations.getImageData</h1>
<p class="desc">Check that exceptions are thrown for operations that are malformed while layers are open.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="200" height="200"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Check that exceptions are thrown for operations that are malformed while layers are open.");
_addTest(function(canvas, ctx) {

  // Shouldn't throw on its own.
  ctx.getImageData(0, 0, 200, 200);
  // Make sure the exception isn't caused by calling the function twice.
  ctx.getImageData(0, 0, 200, 200);
  // Calling again inside a layer should throw.
  ctx.beginLayer();
  assert_throws_dom("InvalidStateError",
                    () => ctx.getImageData(0, 0, 200, 200));

});
</script>
